FEAT: Filetype detection!
Very basic but the editor can detect file types. Another error that is easy to solve, but I don't have time right now: Editor failed to open afile when I pass it as a param, just append a line. Do not create the file, it will be created when we save, as long as the filename is stored in the editor state.